+++ crates/actions/src/operations/mod.rs
@@ -1,5 +1,7 @@
+mod generate_json_schemas;
 mod sync_codeowners;
 mod sync_vcs_hooks;
 
+pub use generate_json_schemas::*;
 pub use sync_codeowners::*;
 pub use sync_vcs_hooks::*;